server/requirements.txt
google-genai==0.3.0
fastapi==0.115.0
uvicorn[standard]==0.30.0
flask==3.0.0
flask-socketio==5.3.0
python-socketio[client]==5.10.0
opencv-python==4.9.0.80
numpy==1.26.0
pillow==10.0.0
pydub==0.25.1
simple-websocket==1.0.0
eventlet==0.35.0
face-recognition==1.3.0   # NEW - For facial recognition